requirements.txt
torch
torchvision
numpy
pandas
matplotlib
pyzmq
certifi
pillow
packaging
transformers
datasets
